Fix merge conflict - move kit to instrument folder
