+++ src/deluge/gui/ui/audio_recorder.cpp
@@ -33,7 +33,7 @@
 #include "model/action/action_logger.h"
 #include "model/clip/instrument_clip.h"
 #include "model/clip/instrument_clip_minder.h"
-#include "model/drum/kit.h"
+#include "model/instrument/kit.h"
 #include "model/sample/sample_recorder.h"
 #include "model/song/song.h"
 #include "playback/playback_handler.h"
