+++ src/deluge/model/instrument/kit.h
